Initial commit, general main menu layout
Still more options to be added to main menu, really just the category
select and adding a min/max to the question count spinner.

It seems the way I'll be going at the moment is to have something of an
SPA written in plain JavaScript, and choosing current page by adding a
`visible` class to children of a pages div. This should be sufficient
enough for this project as there is not going to be many pages or
content. Had it been otherwise, I may have reconsidered.

The biggest part of this at the moment is the styling for the inputs as
to make them larger and more consistent cross-platform. I'm not fully
satisfied with what I have at the moment for that, but it'll do.

The difficulty could ideally be using a fieldset as that is more
semantic and should help accessibility, but styling its legends is more
annoying. I use the ARIA attributes to compensate.